feat(client-reqs): Improve features structure
