+++ client-request-tests/Cargo.toml
@@ -11,11 +11,12 @@ homepage = "https://github.com/iai-callgrind/iai-callgrind"
 repository = "https://github.com/iai-callgrind/iai-callgrind"
 license = "Apache-2.0 OR MIT"
 
-[build-dependencies]
-cc = { version = "1" }
-
 [dependencies]
-cty = { version = "0.2" }
-iai-callgrind = { path = "../iai-callgrind/", version = "=0.9.0", features = [
+iai-callgrind = { path = "../iai-callgrind/", version = "=0.9.0", default-features = false, features = [
   "client_requests_defs",
+] }
+
+[dev-dependencies]
+iai-callgrind = { path = "../iai-callgrind/", version = "=0.9.0", default-features = false, features = [
+  "client_requests",
 ] }
