+++ iai-callgrind/Cargo.toml
@@ -16,9 +16,14 @@ keywords = ["iai", "benchmark", "callgrind", "valgrind"]
 categories = ["development-tools::profiling", "development-tools::testing"]
 
 [features]
+benchmark = [
+  "dep:bincode",
+  "dep:iai-callgrind-macros",
+  "dep:iai-callgrind-runner",
+]
 client_requests = ["client_requests_defs"]
 client_requests_defs = ["dep:cty", "dep:cfg-if"]
-default = []
+default = ["benchmark"]
 ui_tests = []
 
 [build-dependencies]
@@ -29,13 +34,13 @@ bindgen = { version = "0.69" }
 rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
-bincode = "1"
+bincode = { version = "1", optional = true }
 cfg-if = { version = "1", optional = true }
 cty = { version = "0.2", optional = true }
-iai-callgrind-macros = { path = "../iai-callgrind-macros/", version = "=0.1.0" }
+iai-callgrind-macros = { path = "../iai-callgrind-macros/", version = "=0.1.0", optional = true }
 iai-callgrind-runner = { path = "../iai-callgrind-runner/", version = "=0.9.0", default-features = false, features = [
   "api",
-] }
+], optional = true }
 
 [dev-dependencies]
 trybuild = "1"
